refactor: adopt MVC architecture
