+++ controllers/signalController.js
@@ -1,0 +1,21 @@
+import db from "../db.js";
+import { getSignalHistory } from "../scanner.js";
+import { logError } from "../logger.js";
+
+export async function listSignals(req, res) {
+  try {
+    const signals = await db
+      .collection("signals")
+      .find({})
+      .sort({ generatedAt: -1 })
+      .toArray();
+    res.json({ status: "success", signals });
+  } catch (err) {
+    logError("fetching signals", err);
+    res.status(500).json({ error: "Failed to fetch signals" });
+  }
+}
+
+export function signalHistory(req, res) {
+  res.json(getSignalHistory());
+}
